Improvements on parallel search.

+++ src/include/uo_search.h
@@ -38,7 +38,7 @@ extern "C"
 
   typedef struct uo_search_info
   {
-    uint64_t nodes;
+    size_t nodes;
     uo_time time_start;
     uint16_t multipv;
     uint8_t depth;
